Further split the test code behind build tag
