+++ integration_tests/intergamm-osmosis/osmosis.yml
@@ -32,7 +32,6 @@ genesis:
       host_genesis_state:
         params:
           allow_messages: [
-            "/ibc.applications.transfer.v1.MsgTransfer",
             "/osmosis.gamm.v1beta1.MsgCreateBalancerPool",
             "/osmosis.gamm.v1beta1.MsgJoinPool",
             "/osmosis.gamm.v1beta1.MsgExitPool",
